pyproject.toml
[tool.poetry]
name = "catdd"
version = "0.1.0"
description = ""
authors = ["miyashita64 <javajavawookjavawook@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
